test/myCompany/requirements.txt
autopep8==1.5.4
flake8==3.8.4
mccabe==0.6.1
mysql-connector-python==8.0.22
pkg-resources==0.0.0
protobuf==3.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
six==1.15.0
toml==0.10.1
